+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.3.0 (May 26, 2025)
 
 - Requires PHP 8.3+
+- Fixed "empty name" error on invalid file import (Issue #17)
 - Fixed consecutive file imports (Issue #14)
 - Fixed replace file import (Issue #15)